linux sysfs: retrieve vendor, device... info via separate sysfs files
Currently the kernel does not expose the revision file. With that about
to change (due in 4.10) we can read all the information required from
separate files and avoid opening the config one.

The latter has the [negative] side effect of waking up the device, which
in some cases can be quite costly.

Reviewed-by: Adam Jackson <ajax@redhat.com>
Signed-off-by: Emil Velikov <emil.velikov@collabora.com>