configure.ac: remove AC_CONFIG_SRCDIR macro
It is used to guide people who incorrectly set configure --srcdir.

To be actually useful the macro should point to a rather unique file -
which in our current case is a false premise. We have five Makefile.am
in-tree and chances are that $libpciaccess/.. may also contain such a
file.

Let's not pretend that its useful and leave the user to what they're
asking for.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>