configure.ac: set AC_CONFIG_AUX_DIR
If not set, libtool will search directories up to ../.. for an install-sh and
then dump the aux files there. This caused a couple of problems with the xorg
release.sh script that now uses worktrees but is generally bad behaviour
because we can't guarantee that we're not inside some other repository.

Set AC_CONFIG_AUX_DIR to avoid this behavior.

See https://lists.freedesktop.org/archives/xorg-devel/2017-March/053006.html

Note: the commit and above message are shamelessly copied from libinput.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Emil Velikov <emil.velikov@collabora.com>