freebsd_pci: update legacy I/O for new architectures
Update FreeBSD legacy I/O routines for new architectures.  When we're
not on amd64 or i386, and PCI_MAGIC_IO_RANGE is not defined, use IOCTLs
to handle reads and writes through /dev/io.
This is accomplished by opening /dev/io and using IOCTLs from
dev/io/iodev.h, an interface that exists on most FreeBSD architectures.

This change has been in the FreeBSD ports tree for quite some time.