freebsd: Add !legacy open_device_io implementation
Some drivers, such as QXL, rely on this method and do not check for errors
when it is unavailable.  FreeBSD's legacy method can enumerate any d/b/s/f
just fine, so it should be adequate for the !legacy API, as far as I can
tell.

With this change, QXL doesn't crash on startup on FreeBSD.

Signed-off-by: Conrad Meyer <cem@FreeBSD.org>