Obtain correct value of is_64 and is_prefetchable PCI device fields
Correct setting of the is_64 and is_prefetchable pci_device fields in
pci_device_linux_sysfs_probe().
The pci_device struct defines is_64 and is_prefetchable as single bits,
but the previous code was attempting to store the result of a bit-masked
field in a single bit which always resulted in is_64 and is_prefetchable
being zero regardless of the actual capabilities of the PCI device.

Fixes: #15

Signed-off-by: Chester Gillon <chester.gillon@metronet.co.uk>