hurd: Fix pci_device_hurd_map_legacy
It was not passing a proper region number to pci_device_hurd_map_range,
and that would not make sense anyway since the rom is not a region for
instance, and the video memory, interrupt vector etc. aren't a region or
the rom.

So this uses pci_device_hurd_map_range for the rom, and
pci_system_x86_map_dev_mem for non-rom. Unfortunately pci-arbiter
currently cannot get the rom_base from libpciaccess, so we can only
guess that we are trying to map a rom.